+++ example/storybook-nativewind/.ondevice/main.js
@@ -1,0 +1,9 @@
+module.exports = {
+  stories: ['../components/primitives/Button/*.stories.?(ts|tsx|js|jsx)'],
+  addons: [
+    '@storybook/addon-ondevice-notes',
+    '@storybook/addon-ondevice-controls',
+    '@storybook/addon-ondevice-backgrounds',
+    '@storybook/addon-ondevice-actions',
+  ],
+};