+++ example/storybook/.storybook/main.js
@@ -51,13 +51,6 @@ module.exports = {
       ),
     };
 
-    config.resolve.alias = {
-      ...config.resolve.alias,
-      '@gluestack-ui/config': path.join(
-        __dirname,
-        '../../../packages/config/src/gluestack-ui.config'
-      ),
-    };
     config.module.rules.push({
       test: /\.mjs$/,
       include: /node_modules/,