+++ packages/config-v2/src/theme/AlertDialogHeader.ts
@@ -1,0 +1,9 @@
+import { createStyle } from '@gluestack-style/react';
+
+export const AlertDialogHeader = createStyle({
+  p: '$4',
+  borderColor: '$border300',
+  justifyContent: 'space-between',
+  alignItems: 'center',
+  flexDirection: 'row',
+});