+++ packages/config-v2/src/theme/SelectActionsheetItemText.ts
@@ -1,0 +1,10 @@
+import { createStyle } from '@gluestack-style/react';
+
+export const SelectActionsheetItemText = createStyle({
+  mx: '$2',
+  fontSize: '$md',
+  fontFamily: '$body',
+  fontWeight: '$normal',
+  lineHeight: '$md',
+  color: '$text700',
+});