+++ packages/styled/react/src/styled.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
 /* eslint-disable react-hooks/exhaustive-deps */
 import React, {
   ForwardRefExoticComponent,
@@ -97,7 +96,7 @@ function flattenObject(obj: any = {}) {
 function convertUtiltiyToSXFromProps(
   componentProps: any,
   styledSystemProps: any,
-  componentStyleConfig: IComponentStyleConfig,
+  componentStyleConfig: IComponentStyleConfig & { uniqueComponentId: string },
   reservedKeys: any = _reservedKeys,
   plugins: any[] = [],
   ignoreKeys: Set<any> = new Set(),
@@ -120,17 +119,19 @@ function convertUtiltiyToSXFromProps(
 
   if (plugins) {
     for (const pluginName in plugins) {
-      // @ts-ignore
-      [resolvedSxVerbose, , , , sxIgnoreKeys] = plugins[
-        pluginName
-      ]?.inputMiddleWare(
-        resolvedSxVerbose,
-        true,
-        false,
-        Component,
-        componentStyleConfig,
-        ExtendedConfig
-      );
+      if (plugins[pluginName]?.inputMiddleWare) {
+        // @ts-ignore
+        [resolvedSxVerbose, , , , sxIgnoreKeys] = plugins[
+          pluginName
+        ]?.inputMiddleWare(
+          resolvedSxVerbose,
+          true,
+          false,
+          Component,
+          componentStyleConfig,
+          ExtendedConfig
+        );
+      }
     }
 
     sxIgnoreKeys?.forEach((element) => {
@@ -892,6 +893,8 @@ export function verboseStyled<P, Variants, ComCon>(
     descendant: StyleIds;
   };
   let orderedCSSIds: any = [];
+  //@ts-ignore
+  const isStyledComponent = Component?.isStyledComponent;
   // const orderedUnResolvedTheme = updateOrderUnResolvedMap(
   //   theme,
   //   componentHash,
@@ -912,7 +915,10 @@ export function verboseStyled<P, Variants, ComCon>(
       theme,
       componentHash,
       declarationType,
-      componentStyleConfig
+      componentStyleConfig,
+      GluestackStyleSheet,
+      Platform.OS,
+      isStyledComponent
     );
 
     orderedCSSIds = g;
@@ -935,6 +941,17 @@ export function verboseStyled<P, Variants, ComCon>(
     const memoizationKey = sxHash + type;
     // Check if the result is already in the cache
     if (sxMemoizationCache[memoizationKey]) {
+      injectComponentAndDescendantStyles(
+        sxMemoizationCache[memoizationKey],
+        sxHash,
+        type,
+        GluestackStyleSheet,
+        Platform.OS,
+        inlineStyleMap,
+        ignoreKeys,
+        CONFIG
+      );
+
       return sxMemoizationCache[memoizationKey];
     }
 
@@ -1019,6 +1036,8 @@ export function verboseStyled<P, Variants, ComCon>(
     componentStyleConfig?.descendantStyle &&
     componentStyleConfig?.descendantStyle?.length > 0;
 
+  let uniqueComponentId = '';
+
   const StyledComponent = (
     {
       //@ts-ignore
@@ -1034,6 +1053,10 @@ export function verboseStyled<P, Variants, ComCon>(
     ref: React.ForwardedRef<P>
   ) => {
     const isClient = React.useRef(false);
+    const GluestackComponent = useRef(Component);
+    if (uniqueComponentId === '') {
+      uniqueComponentId = componentHash;
+    }
 
     let ignoreKeys: Set<any> = new Set();
 
@@ -1067,6 +1090,8 @@ export function verboseStyled<P, Variants, ComCon>(
       : get();
 
     if (!styleHashCreated) {
+      // eslint-disable-next-line react-hooks/rules-of-hooks
+
       CONFIG = {
         ...styledContext.config,
         propertyTokenMap,
@@ -1106,17 +1131,19 @@ export function verboseStyled<P, Variants, ComCon>(
       if (plugins) {
         for (const pluginName in plugins) {
           let themeIgnoreKeys = new Set();
-          // @ts-ignore
-          [nonVerbosedTheme, , , , themeIgnoreKeys] = plugins[
-            pluginName
-          ]?.inputMiddleWare<P>(
-            nonVerbosedTheme,
-            true,
-            true,
-            componentProps?.as ?? Component,
-            componentStyleConfig,
-            ExtendedConfig
-          );
+          if (plugins[pluginName]?.inputMiddleWare) {
+            // @ts-ignore
+            [nonVerbosedTheme, , , , themeIgnoreKeys] = plugins[
+              pluginName
+            ]?.inputMiddleWare<P>(
+              nonVerbosedTheme,
+              true,
+              true,
+              componentProps?.as ?? Component,
+              { ...componentStyleConfig, uniqueComponentId },
+              ExtendedConfig
+            );
+          }
           themeIgnoreKeys?.forEach((ele) => {
             ignoreKeys.add(ele);
           });
@@ -1470,7 +1497,7 @@ export function verboseStyled<P, Variants, ComCon>(
         //   defaultThemePropsWithoutVariants,
         inlineComponentPropsWithoutVariants,
         styledSystemProps,
-        componentStyleConfig,
+        { ...componentStyleConfig, uniqueComponentId },
         reservedKeys,
         plugins,
         ignoreKeys,
@@ -1487,7 +1514,7 @@ export function verboseStyled<P, Variants, ComCon>(
       convertUtiltiyToSXFromProps(
         mergedPassingProps,
         styledSystemProps,
-        componentStyleConfig,
+        { ...componentStyleConfig, uniqueComponentId },
         reservedKeys,
         plugins,
         ignoreKeys,
@@ -1708,7 +1735,7 @@ export function verboseStyled<P, Variants, ComCon>(
         } = convertUtiltiyToSXFromProps(
           passingPropsUpdated,
           styledSystemProps,
-          componentStyleConfig,
+          { ...componentStyleConfig, uniqueComponentId },
           reservedKeys,
           plugins,
           ignoreKeys,
@@ -2031,7 +2058,10 @@ export function verboseStyled<P, Variants, ComCon>(
               AsComp = plugins[pluginName]?.componentMiddleWare({
                 Component: AsComp,
                 theme,
-                componentStyleConfig,
+                componentStyleConfig: {
+                  ...componentStyleConfig,
+                  uniqueComponentId,
+                },
                 ExtendedConfig,
                 styleCSSIds,
                 GluestackStyleSheet,
@@ -2045,23 +2075,31 @@ export function verboseStyled<P, Variants, ComCon>(
             // @ts-ignore
             if (plugins[pluginName]?.componentMiddleWare) {
               // @ts-ignore
-              Component = plugins[pluginName]?.componentMiddleWare({
-                Component: Component,
+              GluestackComponent.current = plugins[
+                pluginName
+              ]?.componentMiddleWare({
+                Component: GluestackComponent.current,
                 theme,
-                componentStyleConfig,
+                componentStyleConfig: {
+                  ...componentStyleConfig,
+                  uniqueComponentId,
+                },
                 ExtendedConfig,
                 styleCSSIds,
                 GluestackStyleSheet,
               });
 
-              //@ts-ignore
-              pluginData = { ...pluginData, ...Component?.styled };
+              pluginData = {
+                ...pluginData,
+                //@ts-ignore
+                ...GluestackComponent?.current?.styled,
+              };
             }
           }
         }
       }
       return {
-        Component: Component,
+        Component: GluestackComponent.current,
         AsComp: AsComp,
       };
     }, [AsComp]);