+++ packages/unstyled/popover/src/PopoverCloseButton.tsx
@@ -47,6 +47,14 @@ const PopoverCloseButton = (StyledPopoverCloseButton: any) =>
             disabled: isDisabled,
             focusVisible: isFocusVisibleProp || isFocusVisible,
           }}
+          dataSet={{
+            hover: isHoveredProp || isHovered ? 'true' : 'false',
+            focus: isFocusedProp || isFocused ? 'true' : 'false',
+            active: isPressedProp || isPressed ? 'true' : 'false',
+            disabled: isDisabled ? 'true' : 'false',
+            focusVisible:
+              isFocusVisibleProp || isFocusVisible ? 'true' : 'false',
+          }}
           ref={ref}
           disabled={isDisabled}
           onPressIn={composeEventHandlers(onPressIn, pressProps.onPressIn)}