+++ packages/unstyled/popover/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @gluestack-ui/popover
 
+## 0.1.31
+
+### Patch Changes
+
+- fixes
+
+## 0.1.30
+
+### Patch Changes
+
+- Feature
+  Added Support of data-state-\* props
+
+## 0.1.29
+
+### Patch Changes
+
+- - Popover arrow support [PR](https://github.com/gluestack/gluestack-ui/pull/1677)
+  - Popover animation improvements
+
+  - Popover top placement position for mobile
+
+- Updated dependencies
+  - @react-native-aria/overlays@0.3.11
+
 ## 0.1.28
 
 ### Patch Changes