+++ packages/unstyled/select/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @gluestack-ui/select
 
+## 0.1.24
+
+### Patch Changes
+
+- fixes
+
+## 0.1.23
+
+### Patch Changes
+
+- Feature
+  Added Support of data-state-\* props
+- Updated dependencies
+  - @gluestack-ui/form-control@0.1.16
+
+## 0.1.22
+
+### Patch Changes
+
+- `Select` deprecated `selectedLabel`. use `initialLabel` instead.
+
 ## 0.1.21
 
 ### Patch Changes